refactor(RandomNumbers, test): RandomNumbers의 주 생성자를 private으로 변경 & 테스트 패키지 구조 변경
RandomNumbers의 주 생성자를 private으로 변경하였습니다.
또한, 테스트 패키지와 소스 코드 패키지가 동기화 되지 않아 동기화를 시켜주었습니다.
