+++ src/test/java/baseball/domain/SingleNumberTest.java
@@ -1,6 +1,5 @@
-package baseball.domain.number;
+package baseball.domain;
 
-import baseball.domain.SingleNumber;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
