+++ src/test/java/baseball/domain/inputnumber/InputNumberFactoryTest.java
@@ -1,4 +1,4 @@
-package baseball.domain.number.inputnumber;
+package baseball.domain.inputnumber;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
